@@ -1,12 +1,15 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+import typing
+
+from ..player import Player
 from .game import Game
 
 
 class Sandie(Game):
-    def __init__(self, *args, **kwargs):
-        super(Sandie, self).__init__("Sandie", *args, **kwargs)
+    def __init__(self, players: typing.List[Player]) -> None:
+        super().__init__("Sandie", players)
 
 
 # vim: filetype=python
