@@ -1,12 +1,15 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+import typing
+
+from ..player import Player
 from .game import Game
 
 
 class Stroke(Game):
-    def __init__(self, *args, **kwargs):
-        super(Stroke, self).__init__("Stroke", *args, **kwargs)
+    def __init__(self, players: typing.List[Player]) -> None:
+        super().__init__("Stroke", players)
 
 
 # vim: filetype=python
