@@ -2,49 +2,51 @@
 # -*- coding: utf-8 -*-
 
 import abc
+import typing
+
+from ..player import Player
 
 
 class Game(object):
 
     __metaclass__ = abc.ABCMeta
 
-    def __init__(self, name, players, *args, **kwargs):
-        super(Game, self).__init__(*args, **kwargs)
+    def __init__(self, name: str, players: typing.List[Player]) -> None:
+        super(Game, self).__init__()
         self._name = name
         self._players = {}
         for player in players:
             self._players[player] = player
 
     @property
-    def name(self):
+    def name(self) -> str:
         return self._name
 
     @property
-    def players(self):
+    def players(self) -> typing.ValuesView[Player]:
         return self._players.values()
 
-    @property
-    def is_participating(self, player):
+    def is_participating(self, player: Player) -> bool:
         return player in self._players.values()
 
     @abc.abstractmethod
-    def pre_round(self):
+    def pre_round(self) -> None:
         raise NotImplementedError
 
     @abc.abstractmethod
-    def pre_hole(self):
+    def pre_hole(self) -> None:
         raise NotImplementedError
 
     @abc.abstractmethod
-    def at_turn(self):
+    def at_turn(self) -> None:
         raise NotImplementedError
 
     @abc.abstractmethod
-    def post_hole(self):
+    def post_hole(self) -> None:
         raise NotImplementedError
 
     @abc.abstractmethod
-    def post_round(self):
+    def post_round(self) -> None:
         raise NotImplementedError
 
 
